+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 
 
 ## [Unreleased]
+### Added
+* CloneWithAssignmentAttribute can now be used to specify that objects can be cloned with basic assignment, speeding up array cloning considerably.
+
 ### Improved
 * Major revamp of Recorder.Clone functionality to dramatically improve performance (approx 200x in one real-life test case.)